refactor(rate-of-closure): split kinetics core
